backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-bigquery==3.11.4
pydantic==2.7.0

